レートの上限無効化 小数点を含めたレート

+++ src/qow/framework/logic/system/ExecuteLoop.java
@@ -7,7 +7,7 @@ import qow.framework.logic.system.rule.Rule;
  * 可能な限り一秒間に設定されたレートの回数の処理を実行する
  *
  * @author QOW
- * @version 2022/09/03
+ * @version 2022/09/09
  * @since 1.0.0
  */
 public class ExecuteLoop extends Loop {
@@ -19,7 +19,7 @@ public class ExecuteLoop extends Loop {
      *
      * @param rate レートの初期値
      */
-    public ExecuteLoop(int rate) {
+    public ExecuteLoop(double rate) {
         super(rate);
     }
 
@@ -32,7 +32,7 @@ public class ExecuteLoop extends Loop {
 
     /**
      * 可能な限り一秒間に設定されたレートの回数このメソッドを実行する<br>
-     * 指定時間内に処理が終わらない場合は{@link ExecuteLoop#overTime(long)}が呼び出される
+     * 指定時間内に処理が終わらない場合は{@link ExecuteLoop#overTime(double)}が呼び出される
      */
     public void loop() {
         setRate(MainSystem.executeRate);
@@ -50,7 +50,7 @@ public class ExecuteLoop extends Loop {
      *
      * @param overTime 過ぎた時間のナノ秒
      */
-    public void overTime(long overTime) {
+    public void overTime(double overTime) {
         rule.overTimeExecute(overTime);
     }
